requirements.txt
annotated-types==0.5.0
anyio==3.7.1
exceptiongroup==1.1.2
fastapi==0.100.1
h11==0.14.0
httptools==0.6.0
idna==3.4
pydantic==2.1.1
pydantic_core==2.4.0
python-dotenv==1.0.0
celery==5.3.4
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
sqlalchemy==2.0.19
pydantic_settings==2.0.2
psycopg2-binary==2.9.6
PyJWT==2.6.0
cryptography==39.0.1
alembic==1.11.1
pytest==7.4.0
httpx==0.24.1
python-socketio==5.8.0
gevent==22.10.2
requests==2.31.0
gunicorn==20.1.0
GeoAlchemy2==0.14.1
shapely==2.0.1
aioredis==2.0.1
redis==5.0.0
sentry-sdk==1.30.0
python-multipart==0.0.6
boto3==1.25.4
Pillow==9.5.0
colorama==0.4.6